make getFileSize return a int64_t rather than size_t
It's dangerous to convert a file size to size_t on 32 bit platforms.

We use int64_t rather than uint64_t because in some places file sizes
are stored in a lng, which is signed.  Also, st_size is usually a signed
type anyway, so using a signed type means fewer tricky signedness
conversions.

+++ gdk/gdk_logger.c
@@ -1408,7 +1408,7 @@ log_read_transaction(logger *lg, BAT *ids_to_omit, uint32_t *updated, BUN maxupd
 	bool skip_entry = false;
 	ATOMIC_BASE_TYPE dbg = ATOMIC_GET(&GDKdebug);
 	time_t t0 = 0;
-	size_t fs = 0;
+	int64_t fs = 0;
 
 	(void) maxupdated;	/* only used inside assert() */
